Fix includes for sort test, remove unnecessary method, modify filter test
Former-commit-id: 7235bf7b00d583b6be69da5d73bb41691ef34894

+++ core/src/operator/filter/filter_compiler.cpp
@@ -12,7 +12,7 @@ bool testExpressionEvaluater(Table *table, int32_t index)
     case INT64:
         return *((int64_t*) column->getValue(index)) % 2 == 0;
     case DOUBLE:
-        return *((int32_t*) column->getValue(index)) > 10;
+        return *((double*) column->getValue(index)) > 10;
     default:
         break;
     }
